Ensure that imgui sources are available in OUT_DIR
Exposing the location of third_party location to be the build location
does not work when the build location is temporary. Instead, use
OUT_DIR.

Also now that we include the sources in the output explicitly under
`include`, I put the library under `lib` which is customary.

This is a port of `ensure-headers` branch to 0.8.2 tag.